Add course group mapping (#97)
* add course group mapping

* fix wrong table name in migration

* remove double space

* edit CourseGroup constraint

* add course group to user self view

* add userview test

* Add TODOs

* Fix course group get query

* add test

---------

+++ internal/usergroups/middleware.go
@@ -52,8 +52,8 @@ func InjectUserGroupIntoContext(next http.Handler) http.Handler {
 
 		// Get user_group
 		userGroup := model.UserGroup{
-			UserID:  *userID,
-			GroupID: groupID,
+			UserID:  uint(*userID),
+			GroupID: uint(groupID),
 		}
 		var dbUserGroup model.UserGroup
 		err = db.Where(&userGroup).First(&dbUserGroup).Error
